Add files via upload
Zip file of final project for professor Chia Y. Han's CS-5167 User Interface 1 course. To run, simply unzip the folder to a drive location and then right click the index.html file to open it with Chrome. Note : This interface was designed for Chrome and may not work properly in other browsers.